Remove inline helpers
